[refactor]: move `swarm` mod
